back/requirements.txt


Flask==3.0.3
flask-cors
pg8000==1.29.2
pandas
numpy
requests
python-dotenv==1.0.0